requirements.txt
discord.py[voice]==1.7.3
yarl==1.4.2
youtube-dl==2021.6.6
yt-dlp==2023.03.03
# git+https://github.com/ytdl-patched/yt-dlp.git@master
youtube_search==2.1.0
pynacl==1.3.0
python-dotenv==0.13.0
git+https://github.com/nwakefield-cf/pafy_modified.git@develop#egg=pafy
pytz